Extract Bitstyles.Version
