+++ lib/bitstyles_phoenix.ex
@@ -48,7 +48,7 @@ defmodule BitstylesPhoenix do
 
   ### Bitstyles version
 
-  BitstylesPhoenix will generate classes #{BitstylesPhoenix.Bitstyles.version_string()} of bitstyles.
+  BitstylesPhoenix will generate classes #{BitstylesPhoenix.Bitstyles.Version.version_string()} of bitstyles.
   You can set older versions with a configuration:
 
   ```elixir
